+++ src/main/java/elearning/project/models/User.java
@@ -26,7 +26,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 public class User {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int UserID;
+	private Long UserID;
 
 	private String Name;
