+++ src/main/java/elearning/project/repositories/AssessmentRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import elearning.project.models.Assessment;
 
-public interface AssessmentRepo extends JpaRepository<Assessment, Integer> {
+public interface AssessmentRepo extends JpaRepository<Assessment, Long> {
 
 }
