+++ src/main/java/elearning/project/repositories/EnrollmentRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import elearning.project.models.Enrollment;
 
-public interface EnrollmentRepo extends JpaRepository<Enrollment, Integer> {
+public interface EnrollmentRepo extends JpaRepository<Enrollment, Long> {
 
 }
