+++ src/main/java/elearning/project/repositories/SubmissionRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import elearning.project.models.Submission;
 
-public interface SubmissionRepo extends JpaRepository<Submission, Integer> {
+public interface SubmissionRepo extends JpaRepository<Submission, Long> {
 
 }
